card-41689663 make app-bar a container from a component and enhance componentDidUpdate
